Removendo a complexidade desnecessária do projeto de Acesso a Dados
