+++ src/NossoCalendario.Data/Repository/UsuarioRepository.cs
@@ -1,13 +1,31 @@
-﻿using NossoCalendario.Data.Context;
+﻿using Microsoft.EntityFrameworkCore;
+using NossoCalendario.Data.Context;
 using NossoCalendario.Domain.Entities;
-using NossoCalendario.Domain.Interfaces;
+using System;
+using System.Threading.Tasks;
 
 namespace NossoCalendario.Data.Repository
 {
-    public class UsuarioRepository : BaseRepository<Usuario>, IUsuarioRepository
+    public class UsuarioRepository : IDisposable
     {
-        public UsuarioRepository(NossoCalendarioContext context) : base(context)
+        protected readonly NossoCalendarioContext _context;
+        protected readonly DbSet<Usuario> _dbSet;
+
+        public UsuarioRepository(NossoCalendarioContext context)
         {
+            _context = context;
+            _dbSet = context.Set<Usuario>();
+        }
+
+        public async Task Insert(Usuario usuario)
+        {
+            _dbSet.Add(usuario);
+            await _context.SaveChangesAsync();
+        }
+
+        public void Dispose()
+        {
+            _context?.Dispose();
         }
     }
 }
