--- src/NossoCalendario.Domain/Base/IRepository.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace NossoCalendario.Domain.Base
-{
-    public interface IRepository<TEntity> : IDisposable where TEntity : Entity
-    {
-        IUnitOfWork UnitOfWork { get; }
-        Task Insert(TEntity entity);
-    }
-
-}
