--- src/NossoCalendario.Domain/Base/IUnitOfWork.cs
@@ -1,9 +1,0 @@
-﻿using System.Threading.Tasks;
-
-namespace NossoCalendario.Domain.Base
-{
-    public interface IUnitOfWork
-    {
-        Task<bool> Commit();
-    }
-}
