--- src/NossoCalendario.Domain/Interfaces/IUsuarioRepository.cs
@@ -1,10 +1,0 @@
-﻿using NossoCalendario.Domain.Base;
-using NossoCalendario.Domain.Entities;
-using System.Threading.Tasks;
-
-namespace NossoCalendario.Domain.Interfaces
-{
-    public interface IUsuarioRepository : IRepository<Usuario>
-    {
-    }
-}
