+++ src/NossoCalendario.Webapi/Entensions/DependencyInjectionConfiguration.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
 using NossoCalendario.Data.Repository;
-using NossoCalendario.Domain.Interfaces;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
 
@@ -12,7 +11,7 @@ namespace NossoCalendario.WebApi.Entensions
         public static IServiceCollection ResolveDependencies(this IServiceCollection services)
         {
             services.AddTransient<IConfigureOptions<SwaggerGenOptions>, ConfigureSwaggerOptions>();
-            services.AddTransient<IUsuarioRepository, UsuarioRepository>();
+            services.AddTransient<UsuarioRepository>();
             return services;
         }
     }
